Move JniUtils from gluten-arrow to gluten-core
